+++ src/integrators/vcm/vcm_proc.cpp
@@ -100,8 +100,8 @@ public:
 		ref<Timer> timer = new Timer();
 		for (size_t j = 0; j < m_config.sampleCount || (wu->getTimeout() > 0 && (int)timer->getMilliseconds() < wu->getTimeout()); j++) {
 			if (m_config.initialRadius > 0.0f){
-				float reduceFactor = 1.f / std::pow(float(iteration + 1), 0.5f * (1 - 0.75f/*radiusAlpha*/));
-				radius = std::max(reduceFactor * m_config.initialRadius, 1e-7f);
+				Float reduceFactor = 1.f / std::pow(Float(iteration + 1), (Float)(0.5 * (1 - 0.75/*radiusAlpha*/)));
+				radius = std::max(reduceFactor * m_config.initialRadius, (Float)1e-7);
 				iteration += 8;
 			}
 			m_pathSampler->gatherLightPaths(m_config.useVC, m_config.useVM, radius, hilbertCurve.getPointCount(), midres);
